harden: r454 findings 8-9 + CI-leg crash - wide-arithmetic fences across the committee-draft Modular path
The soak waves (and the first dispatched Fuzz leg) kept finding i32
arithmetic on hostile extremes in the legacy committee-draft Modular
decode: the channel-header Varint forms (m1 + 1, 1 - v0, min + span,
now i128 + range check), the BEGABRAC signed-range negation at
lower == i32::MIN (i64 negate), the Gradient predictor sum and the
predicted + diff pixel write (i64 + clamp). All behaviour-preserving
for in-range streams - the 18181-3 conformance compositions stay
bit-exact - and each pinned with its minimized artifact.
decode_modular now soaks clean (26.5M execs).

+++ src/begabrac.rs
@@ -87,7 +87,14 @@ impl Begabrac {
         } else {
             sign = 1;
         }
-        let max = if sign == 1 { upper } else { -lower } as i64;
+        // Hostile-input fence (r454 fuzz): `lower` can be i32::MIN on a
+        // corrupt header, whose i32 negation overflows (debug panic).
+        // Negate in i64.
+        let max = if sign == 1 {
+            upper as i64
+        } else {
+            -(lower as i64)
+        };
         if max <= 0 {
             // Defensive: range was effectively zero on this side.
             return Err(Error::InvalidData(
